Cast to UINT32 to avoid warnings on Windows in ZapBaseRelocs::WriteReloc in src/zap/zaprelocs.cpp

+++ src/zap/zaprelocs.cpp
@@ -92,7 +92,7 @@ void ZapBaseRelocs::WriteReloc(PVOID pSrc, int offset, ZapNode * pTarget, int ta
             // description of IMAGE_REL_BASED_REL_THUMB_MOV32_PCREL
             const UINT32 offsetCorrection = 12;
 
-            UINT32 imm32 = pActualTarget - (pSite + offsetCorrection);
+            UINT32 imm32 = UINT32(pActualTarget - (pSite + offsetCorrection));
 
             PutThumb2Mov32((UINT16 *)pLocation, imm32);
